lesson5/requirements.txt
# Core A2A and agent libraries
a2a-sdk[http-server,grpc,telemetry]>=0.1.0
claude-code-sdk
openai
anthropic

# LiveKit for voice agents
livekit
livekit-agents
livekit-plugins-openai
livekit-plugins-noise-cancellation~=0.2

# Async and networking
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
grpcio>=1.50.0
httpx>=0.24.0
aiofiles>=23.0.0

# Data and ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
pydantic>=2.0.0

# Development tools
python-dotenv>=1.0.0
black>=23.0.0
pylint>=2.17.0
ipython>=8.0.0

# Additional utilities
pyyaml>=6.0
click>=8.1.0
rich>=13.0.0